% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGpath.R
\name{KEGGpath}
\alias{KEGGpath}
\title{Main function}
\usage{
KEGGpath(input, species, n)
}
\arguments{
\item{input}{A txt file with two columns: one with the Entrez Gene ID and another with the tag UP or DOWN.}
}
\description{
Main function
}
\examples{
:
species <- "hsa"
n <- 10
data("de_df")
results <- KEGGpath(de_df, species, n)

}
